Major Change: - Added support to generate user id for new logins - Added support to store userId along with accounts, rules and transactions. - Added support to save transactions in draft state - Added support to query transactions using date filter

+++ routes/v1/account_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, request
+from flask import Blueprint, request, session
 from model.mongodb import accounts_collection, ObjectId
 
 account_routes_bp = Blueprint("account_routes", __name__)
@@ -6,7 +6,8 @@ account_routes_bp = Blueprint("account_routes", __name__)
 
 @account_routes_bp.route("/accounts", methods=["GET"])
 def get_accounts():
-    accounts = list(accounts_collection.find({}))
+    userId = session["user"]["_id"]
+    accounts = list(accounts_collection.find({"userId": userId}))
     for account in accounts:
         account["_id"] = str(account["_id"])
     return {"accounts": accounts}
@@ -14,21 +15,24 @@ def get_accounts():
 
 @account_routes_bp.route("/account", methods=["POST"])
 def upsert_account():
+    userId = session["user"]["_id"]
     data = request.get_json()
     _id = data.pop("_id", None)
     if _id:
-        query = {"_id": ObjectId(_id)}
+        query = {"userId":userId, "_id": ObjectId(_id)}
         result = accounts_collection.update_one(query, {"$set": data})
         data["_id"] = _id
     else:
+        data["userId"] = userId
         result = accounts_collection.insert_one(data)
         data["_id"] = str(result.inserted_id)
-    return data
+    return {"account": data}
 
 
 @account_routes_bp.route("/account/<_id>", methods=["DELETE"])
 def delete_account(_id):
-    result = accounts_collection.delete_one({"_id": ObjectId(_id)})
+    userId = session["user"]["_id"]
+    result = accounts_collection.delete_one({"userId":userId, "_id": ObjectId(_id)})
     if result.deleted_count == 0:
-        return {"error": "Account not found"}, 404
-    return {"message": "Account deleted"}
+        return {"message": "Account not found"}, 404
+    return {"_id": _id}
