+++ routes/v1/rule_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, request
+from flask import Blueprint, request, session
 from model.mongodb import rules_collection, ObjectId
 
 rule_routes_bp = Blueprint("rule_routes", __name__)
@@ -6,7 +6,8 @@ rule_routes_bp = Blueprint("rule_routes", __name__)
 
 @rule_routes_bp.route("/rules", methods=["GET"])
 def get_rules():
-    rules = list(rules_collection.find({}))
+    userId = session["user"]["_id"]
+    rules = list(rules_collection.find({"userId":userId}))
     for rule in rules:
         rule["_id"] = str(rule["_id"])
     return {"rules": rules}
@@ -14,21 +15,24 @@ def get_rules():
 
 @rule_routes_bp.route("/rule", methods=["POST"])
 def upsert_rule():
+    userId = session["user"]["_id"]
     data = request.get_json()
     _id = data.pop("_id", None)
     if _id:
-        query = {"_id": ObjectId(_id)}
+        query = {"userId":userId, "_id": ObjectId(_id)}
         result = rules_collection.update_one(query, {"$set": data})
         data["_id"] = _id
     else:
+        data["userId"] = userId
         result = rules_collection.insert_one(data)
         data["_id"] = str(result.inserted_id)
-    return data
+    return {"rule": data}
 
 
 @rule_routes_bp.route("/rule/<_id>", methods=["DELETE"])
 def delete_rule(_id):
-    result = rules_collection.delete_one({"_id": ObjectId(_id)})
+    userId = session["user"]["_id"]
+    result = rules_collection.delete_one({"userId":userId, "_id": ObjectId(_id)})
     if result.deleted_count == 0:
-        return {"error": "Account not found"}, 404
-    return {"message": "Account deleted"}
+        return {"message": "Rule not found"}, 404
+    return {"_id": _id}
